requirements.txt
streamlit
wordcloud
emoji
matplotlib
urlextractor